+++ src/common/sp.ads
@@ -2,6 +2,6 @@ package SP
     with Pure
 is
 
-    Version : constant String := "0.0.4";
+    Version : constant String := "0.0.5";
 
 end SP;